chore: organize corpus API admin integration tests
