+++ servers/curated-corpus-api/src/admin/resolvers/queries/ScheduledItem/index.ts
@@ -1,8 +1,8 @@
 import { AuthenticationError } from '@pocket-tools/apollo-utils';
-import { getScheduledItems as dbGetScheduledItems } from '../../../database/queries';
-import { ScheduledItemsResult } from '../../../database/types';
-import { ACCESS_DENIED_ERROR } from '../../../shared/types';
-import { IAdminContext } from '../../context';
+import { getScheduledItems as dbGetScheduledItems } from '../../../../database/queries';
+import { ScheduledItemsResult } from '../../../../database/types';
+import { ACCESS_DENIED_ERROR } from '../../../../shared/types';
+import { IAdminContext } from '../../../context';
 
 /**
  * Retrieves a list of Approved Items that are scheduled to appear on a Scheduled Surface
@@ -11,7 +11,7 @@ import { IAdminContext } from '../../context';
 export async function getScheduledItems(
   parent,
   { filters },
-  context: IAdminContext
+  context: IAdminContext,
 ): Promise<ScheduledItemsResult[]> {
   //check if the user does not have the permissions to access this query
   if (
